requirements.txt
selenium
webdriver-manager
pushbullet.py